[GOHW-5]: Added tests to checkout

See merge request trukan24/homework!6

+++ checkout/internal/api/checkout_v1/delete_from_cart_test.go
@@ -1,3 +1,5 @@
+//+build unit
+
 package checkout_v1
 
 import (
@@ -58,7 +60,7 @@ func TestImplementation_DeleteFromCart(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			mockCartService.EXPECT().
-				DeleteFromCart(ctx, tt.req.GetUser(), tt.req.GetSku(), tt.req.GetCount()).
+				DeleteFromCart(ctx, tt.req.GetUser(), tt.req.GetSku(), uint16(tt.req.GetCount())).
 				Return(tt.mockServiceErr).
 				Times(1)
